requirements.txt
httpx>=0.27.0
rich>=13.7.0
pydantic>=2.6.0
google-generativeai>=0.3.0
duckduckgo-search>=5.0.0
edge-tts>=6.1.10
beautifulsoup4>=4.12.0
requests>=2.31.0